LayoutTests/imported/w3c/web-platform-tests/css/css-values/tree-counting/sibling-index-keyframe-length-value-dynamic-expected.txt


PASS Initially, the sibling-index() is 2 for #target
FAIL Removing a preceding sibling of #target reduces the sibling-index() assert_equals: expected "100px" but got "200px"
